GameState: refactor to emulate original tick routine
It now uses a 6502 emulator to run tick routine from the client implementation.

The 6502 bytecode is much like the client .nes, differences are:
 * No iNES header
 * Reset vector points to a specific initialization routine (setuping things without menus)
 * NMI vector points to a specific tick routine (efectively calling game_tick)
 * IRQ vector is set to a panic exit routine (the game does not use IRQs)

It is not a full NES emulator, just the CPU. It supports banking, of course and
some registers are treated specially to avoid infinite loops (ex: VBI bit is
always set in $2002).

It has a very light special API: writing to $ffff stops emulation. It is done to
return from bytecode-specific routine unambiguously. Writing to
global_game_state also stops emulation and notifies that the game is over.

Benefits over previous implementation:
 * Easier to maintain
   * No need to port ASM to C++
   * GameState class is easier to read (it no more implements a full game)
 * Architecturally avoids client/server inconsistencies
 * Allow to monitor game's code with C++ (sometimes easier than with emulator scripts)

Drawbacks:
 * More CPU intensive (180% CPU time observed)