WIP(server-misspredict) add TODO "send eratums with special antilag"
As the message is a reply to sender, round-trip time is the sender's ping, while
"normal" antilag is computed from the time to transit from client a client to
another (client A's ping / 2 + client B's ping / 2).

When clients have vastly divergent pings. That would mean that an eratum sent
to the client with low ping will be over-predicted (in the futur), while an
eratum sent to the client with high ping would be under-predicted (far in the
past)

This commit just adds a TODO for that. The impact was hard to notice, even
with tooled clients and maximal ping difference (one with 180ms ping, another
with 0ms ping.) Also, it would be a pretty big change as antilag computation
is currently not the responsibility of GameInstance. Refactoring for a fix
hard to evaluate. Maybe we need better tools, or fix worst issues first.