DONE(server-misspredict) Fix poorly shared timeline between clients
Summary of server-misspredict:

The main idea is for the server to re-send gamestate to clients if it had
to rollback its internal timeline. As the rollbacked state has been sent, it
means that we "lied" to a client, sending the new state to all clients ensures
that they resync with the new truth.

Doing it, some parts of GameInstance have been extracted from run loop, and made
more generic. Simplifying logic, allowing for further improvements:
 * Handling all waiting messages before sending the computed state,
   * Avoids sending most of states that will be rollbacked, only the eratum is sent
 * Improve timeline sharing by always computing the state at the end of both inputs histories

This series of patchs should greatly improve gameplay between high ping clients.